requirements.txt
## Web Application
Django==2.0.6
django-model-utils==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
dj-database-url==0.5.0
gunicorn==19.7.1
psycopg2==2.7.4
python-dotenv==0.8.2
social-auth-app-django==2.1.0
social-auth-core==1.7.0
whitenoise==3.3.1

## Utilities
boto3==1.7.35
pytz==2018.4
python-dateutil==2.7.2
requests==2.18.4
six==1.11.0

## Subdependencies
#botocore==1.10.35
#certifi==2018.4.16
#chardet==3.0.4
#defusedxml==0.5.0
#docutils==0.14
#idna==2.6
#jmespath==0.9.3
#oauthlib==2.1.0
#PyJWT==1.6.4
#python3-openid==3.1.0
#requests-oauthlib==1.0.0
#s3transfer==0.1.13
#urllib3==1.22

## Python Environment
#pip==10.0.1
#setuptools==39.2.0
#wheel==0.31.1
